requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-tastypie==0.14.3
gunicorn==20.1.0
idna==3.2
postgres==3.0.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
